refactor: util -> utils 폴더명 변경
